fixes in connections for select areas (no more cache-service), added RouteForward for new adapters to use, added simple authentication
