+++ lib/connection.js
@@ -28,22 +28,6 @@ function Connection() {
 }
 
 function setConfig() {
-/*
-    return ((redisIP === '')
-        ? Service.request('service.cache').get('/info')
-        .then(function (result) {
-            return result.json;
-        })
-        .then(function (info) {
-            //redisIP = info.cacheConfig.ip;
-            //redisPort = info.cacheConfig.port;
-            redisIP = sails.config.redis.host;
-            redisPort = sails.config.redis.port;
-            dbArea = info.bladeServices[Service.getName()];
-            return info;
-        })
-        : Promise.resolve());
-*/
     redisIP = sails.config.redis.host;
     redisPort = sails.config.redis.port;
     dbArea = serviceAreas[Service.getName()];
@@ -130,6 +114,12 @@ Connection.prototype.nextGlobalCounter = function () {
         .then(function (data) {
             cli.end();
             return data;
+        })
+        .catch(function(err) {
+            if (cli) {
+                cli.end();
+            }
+            throw err;
         });
 };
